backend/requirements.txt
# Note:
# Coqui TTS 0.22.0 conflicts with librosa on Python 3.10 because TTS pins
# numpy==1.22.0 there. Use Python 3.11+ for this project.
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9
websockets==13.1
aiofiles==23.2.1
python-dotenv==1.0.1
pydantic==2.10.6
loguru==0.7.2
aiosqlite==0.20.0
google-genai>=1.0.0
httpx==0.27.2
soundfile==0.12.1
librosa==0.10.1
numpy==1.26.4
scipy==1.12.0
APScheduler==3.10.4
